removed some warnings
git-svn-id: http://svn.seqan.de/seqan/trunk@10965 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/extras/apps/variant_comp/variant_comp.h
@@ -538,7 +538,7 @@ int compareIndels(
             }
             ++currentId;
         }            
-        while (currentId < length(refIndels))
+        while (currentId < (int)length(refIndels))
         {
 		    write(fileFN,refIndels[currentId],genomeIDs[refIndels[currentId].genomeId],tagAppend,options);
             ++currentId;
